assertion events moved to separate file
