+++ libs/remix-tests/src/assertionEvents.ts
@@ -1,0 +1,12 @@
+const assertionEvents = [
+    {
+        name: 'AssertionEvent',
+        params: ['bool', 'string']
+    },
+    {
+        name: 'AssertionEventUint',
+        params: ['bool', 'string', 'uint256', 'uint256']
+    }
+]
+
+export default assertionEvents
