+++ libs/remix-tests/tests/examples_0/assert_lesserThan_test.sol
@@ -1,4 +1,3 @@
-
 import "remix_tests.sol"; // this import is automatically injected by Remix.
 
 contract AssertLesserThanTest {
